+++ devtools/client/webconsole/components/message-types/ConsoleApiCall.js
@@ -24,6 +24,7 @@ ConsoleApiCall.propTypes = {
   open: PropTypes.bool,
   serviceContainer: PropTypes.object.isRequired,
   timestampsVisible: PropTypes.bool.isRequired,
+  maybeScrollToBottom: PropTypes.func,
 };
 
 ConsoleApiCall.defaultProps = {
@@ -41,6 +42,7 @@ function ConsoleApiCall(props) {
     repeat,
     pausedExecutionPoint,
     isPaused,
+    maybeScrollToBottom,
   } = props;
   const {
     id: messageId,
@@ -66,6 +68,7 @@ function ConsoleApiCall(props) {
     userProvidedStyles,
     serviceContainer,
     type,
+    maybeScrollToBottom,
   };
 
   if (type === "trace") {
@@ -137,6 +140,7 @@ function ConsoleApiCall(props) {
     timeStamp,
     timestampsVisible,
     parameters,
+    maybeScrollToBottom,
   });
 }
 
@@ -150,6 +154,7 @@ function formatReps(options = {}) {
     serviceContainer,
     userProvidedStyles,
     type,
+    maybeScrollToBottom,
   } = options;
 
   return (
@@ -166,6 +171,7 @@ function formatReps(options = {}) {
         loadedObjectProperties,
         loadedObjectEntries,
         type,
+        maybeScrollToBottom,
       }))
       // Interleave spaces.
       .reduce((arr, v, i) => {